+++ src/main/java/io/github/doocs/im/model/request/SendMsgRequest.java
@@ -24,11 +24,14 @@ public class SendMsgRequest extends GenericRequest {
     @JsonProperty("MsgLifeTime")
     private Integer msgLifeTime;
 
+    /**
+     * 消息序列号，用于标记该条消息，数据范围 [0,4294967295]
+     */
     @JsonProperty("MsgSeq")
-    private Integer msgSeq;
+    private Long msgSeq;
 
     @JsonProperty("MsgRandom")
-    private Integer msgRandom;
+    private Long msgRandom;
 
     @JsonProperty("MsgTimeStamp")
     private Integer msgTimeStamp;
@@ -51,14 +54,14 @@ public class SendMsgRequest extends GenericRequest {
     public SendMsgRequest() {
     }
 
-    public SendMsgRequest(String toAccount, Integer msgRandom, List<TIMMsgElement> msgBody) {
+    public SendMsgRequest(String toAccount, Long msgRandom, List<TIMMsgElement> msgBody) {
         this.toAccount = toAccount;
         this.msgRandom = msgRandom;
         this.msgBody = msgBody;
     }
 
     public SendMsgRequest(Integer syncOtherMachine, String fromAccount, String toAccount, Integer msgLifeTime,
-                          Integer msgSeq, Integer msgRandom, Integer msgTimeStamp, List<String> forbidCallbackControl,
+                          Long msgSeq, Long msgRandom, Integer msgTimeStamp, List<String> forbidCallbackControl,
                           List<String> sendMsgControl, List<TIMMsgElement> msgBody, String cloudCustomData,
                           OfflinePushInfo offlinePushInfo) {
         this.syncOtherMachine = syncOtherMachine;
@@ -126,19 +129,19 @@ public class SendMsgRequest extends GenericRequest {
         this.msgLifeTime = msgLifeTime;
     }
 
-    public Integer getMsgSeq() {
+    public Long getMsgSeq() {
         return msgSeq;
     }
 
-    public void setMsgSeq(Integer msgSeq) {
+    public void setMsgSeq(Long msgSeq) {
         this.msgSeq = msgSeq;
     }
 
-    public Integer getMsgRandom() {
+    public Long getMsgRandom() {
         return msgRandom;
     }
 
-    public void setMsgRandom(Integer msgRandom) {
+    public void setMsgRandom(Long msgRandom) {
         this.msgRandom = msgRandom;
     }
 
@@ -196,8 +199,14 @@ public class SendMsgRequest extends GenericRequest {
         private String fromAccount;
         private String toAccount;
         private Integer msgLifeTime;
-        private Integer msgSeq;
-        private Integer msgRandom;
+        /**
+         * 消息序列号，用于标记该条消息，数据范围 [0,4294967295]
+         */
+        private Long msgSeq;
+        /**
+         * 消息随机数，用于标记该条消息，数据范围 [0,4294967295]
+         */
+        private Long msgRandom;
         private Integer msgTimeStamp;
         private List<String> forbidCallbackControl;
         private List<String> sendMsgControl;
@@ -232,12 +241,12 @@ public class SendMsgRequest extends GenericRequest {
             return this;
         }
 
-        public Builder msgSeq(Integer msgSeq) {
+        public Builder msgSeq(Long msgSeq) {
             this.msgSeq = msgSeq;
             return this;
         }
 
-        public Builder msgRandom(Integer msgRandom) {
+        public Builder msgRandom(Long msgRandom) {
             this.msgRandom = msgRandom;
             return this;
         }
